Simple_form used everywhere and the tests play nicely with it

+++ app/controllers/submissions_controller.rb
@@ -11,7 +11,7 @@ class SubmissionsController < ApplicationController
       flash[:error] = ["Sorry, something went wrong. Please try again",
                       @submission.errors.full_messages.to_sentence]
     end
-    redirect_to circle_path(@submission.circle)
+    redirect_to circle_path(params[:circle_id])
   end
 
   def edit_remix
@@ -21,7 +21,7 @@ class SubmissionsController < ApplicationController
 
   def update
     find_submission
-    if @submission.update_attributes(submission_params)
+    if @submission.update_attributes(submission_params) && @submission.remix_id != nil
       undo_link = view_context.link_to('UNDO', delete_remix_path(id: @submission.circle_id, stem_id: @submission.original.id), method: :delete).html_safe
       flash[:success] = "Success! Remix was successfully submitted - #{undo_link}"
       redirect_to circle_path(@submission.circle)
